gui: move add_qrcode to avoid circular dependency for animated qrs
