+++ src/gui/common.py
@@ -1,15 +1,12 @@
 import lvgl as lv
-import qrcode
 import math
 from micropython import const
 import gc
-from .components import QRCode
 
 PADDING    = const(30)
 BTN_HEIGHT = const(80)
 HOR_RES    = const(480)
 VER_RES    = const(800)
-QR_PADDING = const(40)
 
 styles = {}
 
@@ -151,22 +148,6 @@ def add_button_pair(text1, callback1, text2, callback2, scr=None, y=700):
     btn2.set_width(w)
     btn2.set_x(HOR_RES//2+PADDING//2)
     return btn1, btn2
-
-def add_qrcode(text, y=QR_PADDING, scr=None, style=None, width=None):
-    """Helper functions that creates a title-styled label"""
-    if scr is None:
-        scr = lv.scr_act()
-    scr = lv.scr_act()
-
-    if width is None:
-        width = 350
-
-    qr = QRCode(scr)
-    qr.set_text(text)
-    qr.set_size(width)
-    qr.set_text(text)
-    qr.align(scr, lv.ALIGN.IN_TOP_MID, 0, y)
-    return qr
 
 def table_set_mnemonic(table, mnemonic):
     words = mnemonic.split()
